set dtype to int64 to try and fix windows tests

+++ tests/diffraction/filtering_test.py
@@ -30,7 +30,7 @@ def make_data_with_pulse_time(rng, n_event):
     return sc.binning.make_binned(
         events,
         edges=[sc.array(dims=['tof'], values=[10, 500, 1000], unit='us')],
-        groups=[sc.arange('spectrum', 0, 10, unit=None)])
+        groups=[sc.arange('spectrum', 0, 10, unit=None, dtype='int64')])
 
 
 def test_make_data_with_pulse_time():
